data-tramites/pyproject.toml
[project]
name = "tramites"
version = "0.1.0"
description = "Crear DB con los datos de los tramites"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.41",
    "langchain-docling>=0.2.0",
    "langchain-huggingface>=0.1.2",
    "langchain-milvus>=0.1.8",
    "langchain-openai>=0.3.7",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
]
